feat: base logic moving the window with windows key
not in love with it and it seems kinda awkward. Don't know if it its something wrong on how godot treats the windows key or Window it self is eating the input.

Also for some reason the meta_pressed, which should be equivalent to the windows key being pressed doesn't work